Rollback of https://github.com/androidx/media/commit/0605888a9ccabcb0c8fbba122439a6a35045e8c3
*** Original commit ***

Handle int instead of byte in SSIM.

The value of pixels are converted to integers at the point of use,
move this logic to the initialisation step.

This is a prerequisite step for testing SSIM calculation, which
will lead on to some SSIM improvements being verifiable.

Tested manually and SSIM values match for the same video
before and after this change.

***

PiperOrigin-RevId: 473259446
(cherry picked from commit 5092b528d28b925c6b24b1427a75ad2998fb9322)